chore: refactor: move cleanup function to its own file
